requirements.txt
azure-cognitiveservices-speech==1.37.0
blinker==1.8.2
Brotli==1.1.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
filelock==3.14.0
flask==3.0.3
fsspec==2024.5.0
huggingface-hub==0.23.2
idna==3.7
importlib-metadata==7.1.0
intel-openmp==2021.4.0
itsdangerous==2.2.0
jinja2==3.1.4
joblib==1.4.2
MarkupSafe==2.1.5
mkl==2021.4.0
mpmath==1.3.0
mutagen==1.47.0
networkx==3.1
numpy==1.24.4
packaging==24.0
pillow==10.3.0
pycryptodomex==3.20.0
python-dotenv==1.0.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.2
safetensors==0.4.3
scikit-learn==1.3.2
scipy==1.10.1
sentence-transformers==2.7.0
sympy==1.12
tbb==2021.12.0
threadpoolctl==3.5.0
tokenizers==0.19.1
torch==2.3.0
tqdm==4.66.4
transformers==4.41.1
typing-extensions==4.12.0
urllib3==2.2.1
websockets==12.0
werkzeug==3.0.3
yt-dlp==2024.5.26
zipp==3.19.0